move multitarget test to another file
