+++ tests/integration_tests/data_sources/test_clickhouse_ds.py
@@ -3,7 +3,7 @@ import random
 import string
 import unittest
 import requests
-from . import DB_CREDENTIALS, break_dataset
+from . import DB_CREDENTIALS, break_dataset, ClickhouseTest
 from mindsdb_native import Predictor, F
 import gc
 
@@ -19,14 +19,7 @@ def random_string():
     )
 
 
-class TestClickhouse(unittest.TestCase):
-    def setUp(self):
-        self.USER = DB_CREDENTIALS['clickhouse']['user']
-        self.PASSWORD = DB_CREDENTIALS['clickhouse']['password']
-        self.HOST = DB_CREDENTIALS['clickhouse']['host']
-        self.PORT = int(DB_CREDENTIALS['clickhouse']['port'])
-        self.DATABASE = 'test_data'
-
+class TestClickhouse(ClickhouseTest):
     def test_clickhouse_ds(self):
         from mindsdb_native import ClickhouseDS
 
@@ -127,35 +120,3 @@ class TestClickhouse(unittest.TestCase):
             params=params
         )
         assert r.status_code == 200, 'failed to drop temporary database "{}"'.format(TEMP_DB)
-
-    def test_multitarget_prediction(self):
-        from mindsdb_native import ClickhouseDS
-
-
-        LIMIT = 100
-
-        clickhouse_ds = ClickhouseDS(
-            host=self.HOST,
-            port=self.PORT,
-            user=self.USER,
-            password=self.PASSWORD,
-            query='SELECT * FROM {}.{} LIMIT {}'.format(
-                self.DATABASE,
-                'home_rentals',
-                LIMIT
-            )
-        )
-
-        clickhouse_ds.df = break_dataset(clickhouse_ds.df)
-
-        assert len(clickhouse_ds) <= LIMIT
-
-        p = Predictor('test_multitarget_prediction')
-
-        p.learn(
-            from_data=clickhouse_ds,
-            to_predict=['rental_price', 'location'],
-            stop_training_in_x_seconds=3,
-            use_gpu=False,
-            advanced_args={'debug': True}
-        )
